Move csv test into csv submodule
